Multiplicative Mistake: Rename method getSizeOneByOne -> getSize
After finding out how to calculate the size correctly I left only one function to calculate the size of the class
and removed the other one. Now that there is one function to calculate the size, there is no need to distinguish
betwwen the functions and the name can be more generic.